ml-service/requirements.txt
flask
tensorflow
pillow
numpy
flask-cors
opencv-python
gunicorn
scipy
scipy
